requirements.txt
python-dotenv~=1.0.1
google-generativeai
Werkzeug~=3.0.1
gunicorn
bcrypt~=4.1.2
boto3~=1.34.55
pandas~=2.2.1
openpyxl~=3.1.2
requests~=2.31.0
Flask~=3.0.2
flask_cors
langchain~=0.1.12
openai~=1.29.0
beautifulsoup4